+++ source/commands/utils/sharedDangerfileArgs.ts
@@ -28,7 +28,7 @@ export interface SharedCLI extends program.CommanderStatic {
   /** Use GitHub Checks */
   useGithubChecks: boolean
   /** Ignore inline-comments that are in lines which were not changed */
-  noOutOfDiffComments: boolean
+  ignoreOutOfDiffComments: boolean
 }
 
 export default (command: any) =>
@@ -45,4 +45,4 @@ export default (command: any) =>
     .option("-u, --passURLForDSL", "[dev] Use a custom URL to send the Danger DSL into the sub-process")
     .option("-f, --failOnErrors", "Fail if there are fails in the danger report", false)
     .option("--use-github-checks", "Use GitHub Checks", false)
-    .option("--noOutOfDiffComments", "Ignore inline-comments that are in lines which were not changed", false)
+    .option("--ignoreOutOfDiffComments", "Ignore inline-comments that are in lines which were not changed", false)