+++ CHANGELOG.md
@@ -15,8 +15,8 @@
 
 <!-- Your comment below this -->
 
-- Added a CLI option `--noOutOfDiffComments` so that you can ignore inline-comments for lines that were not changed in
-  the checked PR. The comments would be ignored completely - they won't even show in the results comment. [@pinkasey]
+- Added a CLI option `--ignoreOutOfDiffComments` so that you can ignore inline-comments for lines that were not changed
+  in the checked PR. The comments would be ignored completely - they won't even show in the results comment. [@pinkasey]
 
 <!-- Your comment above this -->